refactor(lotto): refactor package structure
