+++ src/main/kotlin/lotto/model/LottoNumber.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.model
 
 data class LottoNumber(val number: Int) {
     init {
