+++ src/test/kotlin/lotto/LottoNumberTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.model.LottoNumber
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
